feat(apex): soql-in-apex support
Adds support for special `_SOQL_` completion item which the Apex LSP server may return
when the cursor is within a SOQL block.  When Apex LSP server returns such item,
"forward" the code-completion request to be handled by the SOQL LSP.

For context: This is the Apex LSP client-side implementation of Approach 2.b described
in https://salesforce.quip.com/NZcqAcKLMCs4.

Pending:
 * Replace out/apex-jorje-lsp.jar with a final signed one
 * Snippets always show up, even on SOQL blocks. One possible solution: move them to Apex LSP server
 * Possible optimization: create the virtual document with only the SOQL block (and then translate
   Positions) instead of replacing everything around the block with blanks
 * Telemetry?